chore: move GetBooksAndRelatedBooksByUserIdRepository to its own file
