+++ src/infra/db/dynamodb/repositories/index.js
@@ -1,5 +1,7 @@
+const GetBooksAndRelatedTasksByUserIdRepository = require('./get-books-and-related-tasks-by-user-id-repository');
 const GetProfileByEmailRepository = require('./get-profile-by-email-repository');
 
 module.exports = {
+  GetBooksAndRelatedTasksByUserIdRepository,
   GetProfileByEmailRepository
 };
